refactor: Split off parameter loading in its own file
